deploy/requirements.txt
distribute
django-extensions
wsgiref
yolk
path.py
South
django-grappelli
fabric
django-compressor
django
pillow
psycopg2
sorl-thumbnail
django-debug-toolbar
ipython
